# [processor/lsminterval] Improve batching performance by using merger.Value