The PR improves the batching performed by the lsminterval processor. With this, there is no need to have batch processor before lsminterval processor.

Related to https://github.com/elastic/opentelemetry-collector-components/issues/245 and https://github.com/elastic/opentelemetry-collector-components/issues/211

<details>
<summary>Benchmark with main using defer merge logic (current PR)</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative                                    20.6µs ± 4%    21.7µs ±14%     ~     (p=0.310 n=5+5)
Aggregation/sum_cumulative-10                                 11.7µs ± 1%    10.2µs ± 7%  -13.24%  (p=0.008 n=5+5)
Aggregation/sum_delta                                         21.6µs ± 4%    22.2µs ± 7%     ~     (p=0.310 n=5+5)
Aggregation/sum_delta-10                                      12.0µs ± 3%     9.9µs ± 2%  -17.67%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative                              20.4µs ± 3%    21.9µs ± 7%   +7.51%  (p=0.032 n=5+5)
Aggregation/histogram_cumulative-10                           12.7µs ± 5%    11.4µs ± 5%   -9.63%  (p=0.008 n=5+5)
Aggregation/histogram_delta                                   22.1µs ±10%    21.6µs ± 5%     ~     (p=0.690 n=5+5)
Aggregation/histogram_delta-10                                12.2µs ± 3%    11.3µs ± 4%   -7.42%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative                           22.6µs ±10%    22.3µs ±14%     ~     (p=0.841 n=5+5)
Aggregation/exphistogram_cumulative-10                        12.8µs ± 5%    11.1µs ± 6%  -12.99%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta                                22.4µs ± 9%    23.1µs ± 9%     ~     (p=0.548 n=5+5)
Aggregation/exphistogram_delta-10                             12.9µs ± 3%    10.8µs ± 2%  -15.86%  (p=0.008 n=5+5)
Aggregation/summary_enabled                                   24.0µs ±11%    24.0µs ± 8%     ~     (p=0.841 n=5+5)
Aggregation/summary_enabled-10                                12.3µs ± 3%    10.4µs ± 3%  -15.26%  (p=0.008 n=5+5)
Aggregation/summary_passthrough                               2.22µs ±16%    2.27µs ±10%     ~     (p=0.548 n=5+5)
Aggregation/summary_passthrough-10                            1.21µs ± 5%    1.12µs ± 6%   -7.30%  (p=0.016 n=5+5)
AggregationWithOTTL/sum_cumulative                            24.6µs ± 1%    24.3µs ± 5%     ~     (p=1.000 n=5+5)
AggregationWithOTTL/sum_cumulative-10                         12.9µs ±13%    12.0µs ± 2%     ~     (p=0.421 n=5+5)
AggregationWithOTTL/sum_delta                                 22.9µs ± 7%    24.3µs ± 5%     ~     (p=0.056 n=5+5)
AggregationWithOTTL/sum_delta-10                              12.8µs ± 2%    11.9µs ± 4%   -6.98%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative                      23.4µs ± 8%    26.4µs ± 5%  +12.85%  (p=0.016 n=5+5)
AggregationWithOTTL/histogram_cumulative-10                   12.9µs ± 2%    12.8µs ± 3%     ~     (p=0.548 n=5+5)
AggregationWithOTTL/histogram_delta                           23.1µs ±12%    25.4µs ± 9%     ~     (p=0.056 n=5+5)
AggregationWithOTTL/histogram_delta-10                        13.4µs ± 7%    12.6µs ± 6%   -6.33%  (p=0.032 n=5+5)
AggregationWithOTTL/exphistogram_cumulative                   24.9µs ±11%    25.7µs ± 5%     ~     (p=0.310 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                14.1µs ± 3%    12.4µs ± 5%  -12.48%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_delta                        24.6µs ± 5%    24.7µs ± 8%     ~     (p=0.421 n=5+5)
AggregationWithOTTL/exphistogram_delta-10                     13.8µs ± 8%    12.6µs ±11%   -8.79%  (p=0.048 n=5+5)
AggregationWithOTTL/summary_enabled                           26.7µs ± 2%    26.4µs ± 7%     ~     (p=0.841 n=5+5)
AggregationWithOTTL/summary_enabled-10                        13.7µs ±14%    12.4µs ± 4%     ~     (p=0.151 n=5+5)
AggregationWithOTTL/summary_passthrough                       2.22µs ± 9%    2.21µs ± 9%     ~     (p=0.841 n=5+5)
AggregationWithOTTL/summary_passthrough-10                    1.26µs ± 4%    1.16µs ± 7%   -7.88%  (p=0.008 n=5+5)

name                                                        old alloc/op   new alloc/op   delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative                                    14.7kB ± 1%    16.2kB ± 1%  +10.10%  (p=0.008 n=5+5)
Aggregation/sum_cumulative-10                                 18.4kB ± 2%    16.6kB ± 1%  -10.03%  (p=0.008 n=5+5)
Aggregation/sum_delta                                         14.6kB ± 1%    16.2kB ± 1%  +10.66%  (p=0.008 n=5+5)
Aggregation/sum_delta-10                                      18.6kB ± 1%    16.4kB ± 1%  -11.52%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative                              17.1kB ± 2%    18.2kB ± 1%   +6.78%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative-10                           21.2kB ± 3%    23.9kB ± 0%  +12.88%  (p=0.016 n=5+4)
Aggregation/histogram_delta                                   17.0kB ± 3%    18.2kB ± 1%   +7.03%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                                21.0kB ± 1%    23.8kB ± 1%  +13.39%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative                           17.2kB ± 1%    18.6kB ± 1%   +8.27%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative-10                        21.6kB ± 2%    19.1kB ± 2%  -11.69%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta                                17.2kB ± 1%    18.6kB ± 0%   +8.24%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta-10                             21.5kB ± 1%    19.1kB ± 1%  -11.53%  (p=0.008 n=5+5)
Aggregation/summary_enabled                                   15.2kB ± 2%    16.4kB ± 2%   +8.07%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                                18.9kB ± 2%    16.8kB ± 1%  -11.13%  (p=0.008 n=5+5)
Aggregation/summary_passthrough                               1.36kB ± 1%    1.54kB ± 0%  +13.58%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                            1.37kB ± 0%    1.54kB ± 1%  +12.04%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_cumulative                            17.3kB ± 1%    18.7kB ± 2%   +8.12%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_cumulative-10                         19.4kB ±12%    19.3kB ± 1%     ~     (p=0.651 n=5+5)
AggregationWithOTTL/sum_delta                                 17.2kB ± 2%    18.4kB ± 1%   +7.24%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_delta-10                              21.6kB ± 0%    19.2kB ± 1%  -11.06%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative                      19.6kB ± 1%    20.8kB ± 2%   +5.92%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative-10                   23.9kB ± 1%    21.3kB ± 1%  -10.81%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta                           19.6kB ± 1%    20.7kB ± 2%   +5.52%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10                        23.1kB ± 5%    21.0kB ± 6%   -9.07%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative                   19.5kB ± 1%    21.1kB ± 1%   +7.95%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                24.3kB ± 6%    21.7kB ± 1%  -10.93%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_delta                        19.6kB ± 1%    21.0kB ± 5%   +7.17%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_delta-10                     24.0kB ± 4%    21.7kB ± 0%   -9.65%  (p=0.016 n=5+4)
AggregationWithOTTL/summary_enabled                           18.0kB ± 2%    19.0kB ± 2%   +5.59%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_enabled-10                        22.1kB ± 1%    19.6kB ± 1%  -11.43%  (p=0.016 n=4+5)
AggregationWithOTTL/summary_passthrough                       1.35kB ± 1%    1.54kB ± 0%  +13.67%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_passthrough-10                    1.37kB ± 1%    1.53kB ± 1%  +11.43%  (p=0.008 n=5+5)

name                                                        old allocs/op  new allocs/op  delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative                                       155 ± 0%       155 ± 1%     ~     (p=1.000 n=5+5)
Aggregation/sum_cumulative-10                                    199 ± 1%       155 ± 1%  -21.85%  (p=0.008 n=5+5)
Aggregation/sum_delta                                            156 ± 0%       156 ± 1%     ~     (p=1.000 n=5+5)
Aggregation/sum_delta-10                                         201 ± 0%       156 ± 1%  -22.64%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative                                 169 ± 0%       171 ± 1%   +0.95%  (p=0.048 n=5+5)
Aggregation/histogram_cumulative-10                              209 ± 1%       217 ± 0%   +3.83%  (p=0.016 n=5+4)
Aggregation/histogram_delta                                      170 ± 0%       172 ± 1%   +1.06%  (p=0.048 n=5+5)
Aggregation/histogram_delta-10                                   209 ± 1%       217 ± 0%   +3.73%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative                              173 ± 0%       174 ± 1%     ~     (p=0.444 n=5+5)
Aggregation/exphistogram_cumulative-10                           220 ± 1%       175 ± 1%  -20.47%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta                                   174 ± 0%       174 ± 1%     ~     (p=1.000 n=5+5)
Aggregation/exphistogram_delta-10                                220 ± 1%       176 ± 1%  -20.07%  (p=0.008 n=5+5)
Aggregation/summary_enabled                                      164 ± 0%       164 ± 1%     ~     (p=0.444 n=5+5)
Aggregation/summary_enabled-10                                   208 ± 1%       165 ± 1%  -20.60%  (p=0.008 n=5+5)
Aggregation/summary_passthrough                                 35.0 ± 0%      33.0 ± 0%   -5.71%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                              34.0 ± 0%      32.0 ± 0%   -5.88%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_cumulative                               182 ± 0%       181 ± 1%     ~     (p=0.444 n=5+5)
AggregationWithOTTL/sum_cumulative-10                            203 ±16%       182 ± 1%     ~     (p=0.325 n=5+5)
AggregationWithOTTL/sum_delta                                    183 ± 0%       182 ± 0%   -0.55%  (p=0.000 n=5+4)
AggregationWithOTTL/sum_delta-10                                 234 ± 0%       183 ± 1%  -22.10%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative                         196 ± 0%       197 ± 1%     ~     (p=1.000 n=5+5)
AggregationWithOTTL/histogram_cumulative-10                      240 ± 1%       197 ± 1%  -17.61%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta                              197 ± 0%       198 ± 2%   +0.51%  (p=0.048 n=5+5)
AggregationWithOTTL/histogram_delta-10                           244 ± 0%       199 ± 0%  -18.21%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative                      200 ± 0%       201 ± 1%   +0.60%  (p=0.048 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                   255 ± 1%       202 ± 0%  -20.66%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_delta                           201 ± 0%       201 ± 1%     ~     (p=0.762 n=5+5)
AggregationWithOTTL/exphistogram_delta-10                        256 ± 1%       203 ± 0%  -20.77%  (p=0.016 n=5+4)
AggregationWithOTTL/summary_enabled                              191 ± 0%       192 ± 1%     ~     (p=0.444 n=5+5)
AggregationWithOTTL/summary_enabled-10                           243 ± 0%       192 ± 0%  -21.07%  (p=0.000 n=4+5)
AggregationWithOTTL/summary_passthrough                         35.0 ± 0%      33.0 ± 0%   -5.71%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_passthrough-10                      34.0 ± 0%      32.0 ± 0%     ~     (p=0.079 n=4+5)
```

</details>

<details>
<summary>Benchmark diff with main using eager merge logic (previous PR) and serial benchmark</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative-10                                 12.3µs ± 2%    11.5µs ± 3%   -6.02%  (p=0.008 n=5+5)
Aggregation/sum_delta-10                                      12.8µs ± 7%    11.7µs ± 3%   -9.03%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative-10                           14.3µs ± 3%    12.9µs ± 7%  -10.30%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                                14.5µs ± 6%    13.1µs ± 2%   -9.58%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative-10                        13.9µs ± 2%    12.5µs ± 3%   -9.91%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta-10                             14.2µs ± 3%    12.8µs ± 2%   -9.62%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                                13.6µs ± 5%    12.4µs ± 3%   -8.86%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                            1.42µs ± 9%    1.16µs ± 3%  -18.44%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_cumulative-10                         14.2µs ± 4%    13.9µs ± 6%     ~     (p=0.548 n=5+5)
AggregationWithOTTL/sum_delta-10                              14.6µs ± 4%    13.7µs ± 9%     ~     (p=0.095 n=5+5)
AggregationWithOTTL/histogram_cumulative-10                   15.2µs ± 2%    14.9µs ± 3%     ~     (p=0.310 n=5+5)
AggregationWithOTTL/histogram_delta-10                        15.2µs ± 1%    14.1µs ± 3%   -7.20%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                15.2µs ± 8%    14.4µs ± 7%     ~     (p=0.095 n=5+5)
AggregationWithOTTL/exphistogram_delta-10                     15.5µs ± 2%    14.5µs ± 2%   -6.88%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_enabled-10                        15.3µs ± 1%    13.9µs ± 5%   -9.14%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_passthrough-10                    1.41µs ± 3%    1.19µs ± 3%  -15.67%  (p=0.008 n=5+5)
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor/internal/merger goos:darwin goarch:arm64
Merge/without_overflow/sum_cumulative_overflow-10             4.23µs ± 3%    4.35µs ±13%     ~     (p=0.690 n=5+5)
Merge/without_overflow/sum_delta_overflow-10                  4.28µs ± 2%    4.36µs ± 7%     ~     (p=0.222 n=5+5)
Merge/without_overflow/histogram_cumulative_overflow-10       4.84µs ± 5%    4.51µs ± 5%   -6.83%  (p=0.008 n=5+5)
Merge/without_overflow/histogram_delta_overflow-10            4.71µs ± 5%    4.53µs ± 2%     ~     (p=0.095 n=5+5)
Merge/without_overflow/exphistogram_cumulative_overflow-10    4.70µs ± 4%    4.57µs ± 1%     ~     (p=0.095 n=5+5)
Merge/without_overflow/exphistogram_delta_overflow-10         4.58µs ± 1%    4.81µs ± 4%   +4.85%  (p=0.016 n=5+5)
Merge/with_overflow/sum_cumulative_overflow-10                3.61µs ±18%    3.69µs ± 7%     ~     (p=0.421 n=5+5)
Merge/with_overflow/sum_delta_overflow-10                     3.57µs ± 5%    3.40µs ± 2%   -4.84%  (p=0.016 n=5+5)
Merge/with_overflow/histogram_cumulative_overflow-10          3.86µs ± 3%    3.69µs ± 5%     ~     (p=0.095 n=5+5)
Merge/with_overflow/histogram_delta_overflow-10               3.87µs ± 7%    3.67µs ± 1%     ~     (p=0.421 n=5+5)
Merge/with_overflow/exphistogram_cumulative_overflow-10       3.80µs ± 5%    3.84µs ± 6%     ~     (p=1.000 n=5+5)
Merge/with_overflow/exphistogram_delta_overflow-10            4.14µs ± 5%    3.54µs ± 1%  -14.40%  (p=0.008 n=5+5)

name                                                        old alloc/op   new alloc/op   delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative-10                                 13.6kB ± 0%    11.3kB ± 0%  -16.69%  (p=0.016 n=5+4)
Aggregation/sum_delta-10                                      14.4kB ± 9%    11.4kB ± 1%  -21.17%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative-10                           20.4kB ± 1%    18.4kB ± 1%   -9.97%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                                20.6kB ± 1%    18.5kB ± 1%  -10.17%  (p=0.016 n=4+5)
Aggregation/exphistogram_cumulative-10                        16.0kB ± 0%    13.7kB ± 1%  -14.15%  (p=0.016 n=4+5)
Aggregation/exphistogram_delta-10                             16.1kB ± 1%    13.9kB ± 1%  -13.68%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                                14.0kB ± 1%    11.9kB ± 1%  -15.26%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                            1.35kB ± 1%    1.20kB ± 0%  -10.50%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_cumulative-10                         15.9kB ± 1%    14.1kB ± 1%  -11.16%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_delta-10                              15.9kB ± 0%    14.1kB ± 1%  -11.33%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative-10                   18.3kB ± 1%    16.0kB ± 1%  -12.45%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10                        18.2kB ± 1%    15.7kB ± 8%  -13.62%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                18.3kB ± 1%    16.1kB ± 8%  -12.00%  (p=0.016 n=4+5)
AggregationWithOTTL/exphistogram_delta-10                     18.4kB ± 1%    16.5kB ± 1%  -10.11%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_enabled-10                        16.4kB ± 1%    14.5kB ± 1%  -11.60%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_passthrough-10                    1.34kB ± 1%    1.20kB ± 1%  -10.04%  (p=0.008 n=5+5)
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor/internal/merger goos:darwin goarch:arm64
Merge/without_overflow/sum_cumulative_overflow-10             5.05kB ± 0%    5.05kB ± 0%     ~     (p=0.333 n=5+4)
Merge/without_overflow/sum_delta_overflow-10                  5.05kB ± 0%    5.05kB ± 0%     ~     (p=0.167 n=5+5)
Merge/without_overflow/histogram_cumulative_overflow-10       5.58kB ± 0%    5.58kB ± 0%     ~     (all equal)
Merge/without_overflow/histogram_delta_overflow-10            5.58kB ± 0%    5.58kB ± 0%     ~     (all equal)
Merge/without_overflow/exphistogram_cumulative_overflow-10    5.69kB ± 0%    5.69kB ± 0%     ~     (all equal)
Merge/without_overflow/exphistogram_delta_overflow-10         5.71kB ± 0%    5.71kB ± 0%     ~     (all equal)
Merge/with_overflow/sum_cumulative_overflow-10                5.04kB ± 0%    5.04kB ± 0%     ~     (all equal)
Merge/with_overflow/sum_delta_overflow-10                     5.04kB ± 0%    5.04kB ± 0%     ~     (all equal)
Merge/with_overflow/histogram_cumulative_overflow-10          5.22kB ± 0%    5.22kB ± 0%     ~     (all equal)
Merge/with_overflow/histogram_delta_overflow-10               5.22kB ± 0%    5.22kB ± 0%     ~     (all equal)
Merge/with_overflow/exphistogram_cumulative_overflow-10       5.25kB ± 0%    5.25kB ± 0%     ~     (all equal)
Merge/with_overflow/exphistogram_delta_overflow-10            5.27kB ± 0%    5.27kB ± 0%     ~     (all equal)

name                                                        old allocs/op  new allocs/op  delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative-10                                    154 ± 0%       122 ± 0%  -20.78%  (p=0.000 n=5+4)
Aggregation/sum_delta-10                                         155 ± 0%       122 ± 1%  -21.16%  (p=0.008 n=5+5)
…