@lahsivjar 2025-02-06
I have updated the PR to use defer merging of values (the approach @axw PoCd [here](https://github.com/axw/opentelemetry-collector-components/commit/b71bfa9e4fc6c4c5df79a8cfdfbc0e344f393fd7)) even though it is a bit inferior in serial benchmarks to eager merging logic which the PR. This is because the PR shows much better performance in concurrent benchmarks:

<img width="1387" alt="Screenshot 2025-02-07 at 12 29 28 AM" src="https://github.com/user-attachments/assets/c7e86123-fc14-4ecd-8dbc-c6b1ffe59e9d" />