@lahsivjar 2025-02-06
I just realized that I have been running the past few benchmarks without cleaning up the pebble directory which might have skewed the results so I have deleted the last update. I am putting this in draft while I run the benchmarks again and reevaluate the 3 approaches we have:

1. With current main + batch processor
2. With better batching performance using merger.Value but merging eagerly in ConsumeMetrics
3. With better batching performance using merger.Value but defer merging until needed