@lahsivjar 2025-02-17
I have started working on another approach to improve batching by using a method local `merger.Value` to create the initial value when iterating over the input `pmetric.Metrics` and then merging this `merger.Value` to a processor local `merger.Value` with lock. Theoretically, merging a `merger.Value` to another `merger.Value` would be more efficient than the using the `pebble.Batch` approach we had earlier and would also have lesser lock contention than what we have in this PR.