@lahsivjar 2025-03-31
Closing this as we are in a good spot after optimizing exponential histograms and merges are already efficient as a result. While the ideas presented in the PR are have some scope for improving the performance it would be better to approach these with targeted goals and choose the best option.